float changed to double for total_price in ThanksPageResponse.kt

+++ features/payment/thankyou_native/src/main/java/com/tokopedia/thankyou_native/data/mapper/FeatureRecommendationMapper.kt
@@ -63,7 +63,7 @@ object FeatureRecommendationMapper {
                 amount += orderItem.totalPrice
             }
             shopId = it.storeId.toIntOrZero()
-            shopParams.add(MembershipOrderData(shopId, amount))
+            shopParams.add(MembershipOrderData(shopId, amount.toFloat()))
         }
         if (!engineData.featureEngineItem.isNullOrEmpty()) {
             engineData.featureEngineItem.forEachIndexed { i, featureEngineItem ->
